LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/url-encoded-expected.txt



PASS text.simple 
PASS textarea.simple 
FAIL nokeygen.simple assert_equals: expected "foo=barb" but got "foo=barb&=MIICQTCCASkwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDV4hN0dzTdIJjp3SwmYnjqFO91JgKfMCMacefbSNXCoksKHqoBisNVcB%2Fn1XMB0Vk5CjXeTiukRXWdRJObTMBb9Y2%2BtFeyXuRjsoQsZiHp%2F5XfgBC0i9YMT4MpHvGosOu%2BdZAxEYyOgROf%2FaaSmEe4tbcIxSdtBQulbvGK4OhvB1pM1SqCQOrsrwV%2FzEYqt6K2nU8kd43f1cRP9%2FCHlMdBtHHt54%2B3oMfOZqa8SAQ8eL7WWgIrYcIfkQozc8U7xozvnNOZ38jXh6x%2Fh00JPxnGeLsMC6mqYYdPKJO33P6wpKL8AOl0VPeBMSw8fv3AHxp0%2Bk3S%2Bmq4BpSVxQEE8%2FCbAgMBAAEWAQAwDQYJKoZIhvcNAQEEBQADggEBAJl9r9jfDkAKK%2BEV%2BjHd78n%2F5AygR%2Ft2dafQABxjS9H08L99JaHV1S8nmPLEM3IZvXkLDZTR3ML1jPOH0hPO8ecfFo9iFheh3QxmSt%2Bb2r9PFe9V3FCAWgxN1UBI8LJyk%2BdEzU%2FRvyS8BL7syomzaVEEXOmdkOrIYx9u5ikva%2FqQkQ%2BMD2O6oSy%2Fof5Atb427CuxPen9SptCBVxGLf5dbl41gmlyDgygxBZfq2UcLU1C5B6ybVDB8TzcStsVQjGXCx31TM2uuyaLzMQ42w4OkWHQNaD%2BQZAE%2B1gQSs%2FztNRj%2BF4woc1PTEhK8NT%2B0AifMNuAkjHddsYOJeO2zgDSpdY%3D"
